requirements.txt
fastapi
uvicorn
httpx
python-multipart
gemini-srt-translator
python-dotenv
charset-normalizer
pydantic-settings